Add local tool versions
fixes #362

Neil inquired about "a hook to collect the versions of NodeJS/ruby/bundler/compiler present on a user's workstation".  I use asdf to manage all my runtime versions.  It uses the .tool-versions file to pin local versions.  This won't guarantee that all developers are using asdf but if they are this will enforce consistency and track changes.